完善代理逻辑

+++ XrayServer/XrayProcessHelper.cs
@@ -5,10 +5,15 @@ namespace XrayServer
     public class XRayProcessHelper
     {
         private readonly string _xrayPath;
-        private readonly Process _process = new();
-        private bool _isRunning = false;
+        private Process? _process = null;
 
         public Action<string>? OutputLineAction;
+
+        // 用一个常驻的单线程调度xray进程
+        private Scheduler _scheduler = new(1, ThreadPriority.Normal);
+
+        private int _isRuning = 0; 
+
         public XRayProcessHelper()
         {
             var currentPath = Environment.CurrentDirectory;
@@ -17,20 +22,19 @@ namespace XrayServer
 
         public async void Start(string configPath)
         {
-            if (_isRunning) return;
-
-            _isRunning = true;
-
-            await Task.Run(() =>
+            await Task.Factory.StartNew(() =>
             {
                 try
                 {
+                    _process = new Process();
                     _process.StartInfo.UseShellExecute = false;
                     _process.StartInfo.FileName = _xrayPath;
-                    _process.StartInfo.CreateNoWindow= true;
+                    _process.StartInfo.CreateNoWindow = true;
                     _process.StartInfo.Arguments = $"run -config {configPath}";
                     _process.StartInfo.RedirectStandardOutput = true;
                     _process.Start();
+
+                    Interlocked.Increment(ref _isRuning);
 
                     // 利用job作业系统管理子进程的生命周期
                     ChildProcessTracker.AddProcess(_process);
@@ -45,19 +49,20 @@ namespace XrayServer
                     }
 
                     _process.WaitForExit();
+
+                    Interlocked.Decrement(ref _isRuning);
                 }
                 catch (Exception ex)
                 {
                     Debug.WriteLine(ex.Message);
                 }
-            });
-
-            _isRunning = false;
+            }, CancellationToken.None, TaskCreationOptions.None, _scheduler);
         }
 
         public void Stop()
         {
-            _process.Kill();
+            if (Interlocked.CompareExchange(ref _isRuning, 0, 0) != 0 && _process != null)
+                _process.Kill();
         }
     }
 }
